Add support for redirecting to external login URL on 401.
Add "external-login-url" option to UI that, when set, will
cause users to be sent to that URL to log the user back in.
This can be useful with the "remote" auth type so that 401s
returned when making backend requests properly cause the user
to get logged back in instead of just failing.